feat(src): extract quality observation logic into new module
- Refactored quality-related logic into a new module `youtube_quality.rs`.
- Moved QualityObservation, QualityObservedResponse, and QualityObservedItag structs to the new module.
- Updated `YoutubeState` to include pub(crate) access for quality_observations and quality_streams.
- Extracted quality observation handling functions (`get_video_quality_observed`, `get_video_quality_stream`) into the new module.
- Removed deprecated and redundant code from `src/main.rs` and `src/youtube.rs`.
