+++ src/youtube.rs
@@ -4,23 +4,17 @@ use axum::{
     extract::{Path, RawQuery, State},
     http::{HeaderMap, HeaderName, HeaderValue, StatusCode, header},
     middleware,
-    response::{
-        IntoResponse, Response,
-        sse::{Event as SseEvent, KeepAlive, Sse},
-    },
+    response::{IntoResponse, Response},
     routing::get,
 };
-use futures_util::StreamExt;
 use quick_xml::{Reader, Writer, events::Event as XmlEvent};
 use regex::{Captures, Regex};
 use serde::{Deserialize, Serialize};
 use std::{
     collections::HashMap,
     sync::{Arc, Mutex, OnceLock},
-    time::{SystemTime, UNIX_EPOCH},
 };
 use tokio::sync::broadcast;
-use tokio_stream::wrappers::BroadcastStream;
 
 use crate::{
     auth::{self, WebAuthConfig},
@@ -30,6 +24,10 @@ use crate::{
         InvidiousClient, YoutubeChannel, YoutubeChannelInfo, YoutubePlaylist, YoutubeVideo,
         YoutubeVideoMeta, is_valid_video_id,
     },
+    youtube_quality::{
+        QualityObservation, QualityObservedResponse, get_video_quality_observed,
+        get_video_quality_stream, observe_quality_from_request,
+    },
 };
 
 /// State for YouTube routes
@@ -37,8 +35,9 @@ use crate::{
 pub struct YoutubeState {
     invidious: Option<InvidiousClient>,
     invidious_base_url: Option<String>,
-    quality_observations: Arc<Mutex<HashMap<String, QualityObservation>>>,
-    quality_streams: Arc<Mutex<HashMap<String, broadcast::Sender<QualityObservedResponse>>>>,
+    pub(crate) quality_observations: Arc<Mutex<HashMap<String, QualityObservation>>>,
+    pub(crate) quality_streams:
+        Arc<Mutex<HashMap<String, broadcast::Sender<QualityObservedResponse>>>>,
 }
 
 impl YoutubeState {
@@ -93,27 +92,6 @@ pub struct ChannelInfoResponse {
 #[derive(Debug, Serialize)]
 pub struct VideoMetaResponse {
     pub video: YoutubeVideoMeta,
-}
-
-#[derive(Debug, Clone)]
-struct QualityObservation {
-    current_itag: String,
-    current_quality_label: String,
-    counts: HashMap<String, u64>,
-    last_updated_unix_secs: u64,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct QualityObservedResponse {
-    pub current_quality_label: Option<String>,
-    pub seen_itags: Vec<QualityObservedItag>,
-    pub last_updated_unix_secs: Option<u64>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct QualityObservedItag {
-    pub quality_label: String,
-    pub count: u64,
 }
 
 /// Playlist list response
@@ -255,68 +233,6 @@ async fn get_video_meta(
         Ok(video) => (StatusCode::OK, Json(VideoMetaResponse { video })).into_response(),
         Err(e) => e.into_response(),
     }
-}
-
-async fn get_video_quality_observed(
-    State(state): State<YoutubeState>,
-    Path(video_id): Path<String>,
-) -> Response {
-    if !is_valid_video_id(&video_id) {
-        return (StatusCode::BAD_REQUEST, "invalid video_id format").into_response();
-    }
-
-    let observations = match state.quality_observations.lock() {
-        Ok(o) => o,
-        Err(_) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "quality observation store unavailable",
-            )
-                .into_response();
-        }
-    };
-
-    let Some(obs) = observations.get(&video_id) else {
-        return (
-            StatusCode::OK,
-            Json(QualityObservedResponse {
-                current_quality_label: None,
-                seen_itags: Vec::new(),
-                last_updated_unix_secs: None,
-            }),
-        )
-            .into_response();
-    };
-
-    (StatusCode::OK, Json(build_quality_observed_response(obs))).into_response()
-}
-
-async fn get_video_quality_stream(
-    State(state): State<YoutubeState>,
-    Path(video_id): Path<String>,
-) -> Response {
-    if !is_valid_video_id(&video_id) {
-        return (StatusCode::BAD_REQUEST, "invalid video_id format").into_response();
-    }
-
-    let sender = get_or_create_quality_sender(&state, &video_id);
-    let receiver = sender.subscribe();
-    let stream = BroadcastStream::new(receiver).filter_map(|msg| async move {
-        match msg {
-            Ok(update) => serde_json::to_string(&update).ok().map(|json| {
-                Ok::<SseEvent, std::convert::Infallible>(SseEvent::default().data(json))
-            }),
-            Err(_) => None,
-        }
-    });
-
-    Sse::new(stream)
-        .keep_alive(
-            KeepAlive::new()
-                .interval(std::time::Duration::from_secs(25))
-                .text("hb"),
-        )
-        .into_response()
 }
 
 /// Get authenticated user's playlists
@@ -527,7 +443,6 @@ static COMPANION_ROOT_DOUBLE_RE: OnceLock<Regex> = OnceLock::new();
 static COMPANION_ROOT_SINGLE_RE: OnceLock<Regex> = OnceLock::new();
 static STATIC_ROOT_DOUBLE_RE: OnceLock<Regex> = OnceLock::new();
 static STATIC_ROOT_SINGLE_RE: OnceLock<Regex> = OnceLock::new();
-static EMBED_REFERER_VIDEO_ID_RE: OnceLock<Regex> = OnceLock::new();
 
 const COMPANION_PROXY_PREFIX: &str = "/api/youtube/companion/api/";
 const VIDEO_PROXY_PREFIX: &str = "/api/youtube/proxy/";
@@ -812,196 +727,6 @@ fn header_value_string(value: Option<&HeaderValue>) -> String {
         .and_then(|v| v.to_str().ok())
         .unwrap_or("<none>")
         .to_string()
-}
-
-fn build_quality_observed_response(obs: &QualityObservation) -> QualityObservedResponse {
-    let mut seen_itags: Vec<QualityObservedItag> = obs
-        .counts
-        .iter()
-        .map(|(itag, count)| QualityObservedItag {
-            quality_label: itag_quality_label(itag).to_string(),
-            count: *count,
-        })
-        .collect();
-    seen_itags.sort_by(|a, b| {
-        b.count
-            .cmp(&a.count)
-            .then_with(|| a.quality_label.cmp(&b.quality_label))
-    });
-
-    QualityObservedResponse {
-        current_quality_label: Some(obs.current_quality_label.clone()),
-        seen_itags,
-        last_updated_unix_secs: Some(obs.last_updated_unix_secs),
-    }
-}
-
-fn is_video_itag(itag: &str) -> bool {
-    matches!(
-        itag,
-        "160"
-            | "278"
-            | "394"
-            | "133"
-            | "242"
-            | "395"
-            | "134"
-            | "243"
-            | "396"
-            | "18"
-            | "135"
-            | "244"
-            | "397"
-            | "136"
-            | "247"
-            | "398"
-            | "22"
-            | "298"
-            | "137"
-            | "248"
-            | "399"
-            | "299"
-    )
-}
-
-fn get_or_create_quality_sender(
-    state: &YoutubeState,
-    video_id: &str,
-) -> broadcast::Sender<QualityObservedResponse> {
-    let mut streams = match state.quality_streams.lock() {
-        Ok(s) => s,
-        Err(poisoned) => poisoned.into_inner(),
-    };
-
-    streams
-        .entry(video_id.to_string())
-        .or_insert_with(|| {
-            let (tx, _) = broadcast::channel(128);
-            tx
-        })
-        .clone()
-}
-
-fn now_unix_secs() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|d| d.as_secs())
-        .unwrap_or(0)
-}
-
-fn decode_query_value(value: &str) -> Option<String> {
-    urlencoding::decode(value).ok().map(|v| v.into_owned())
-}
-
-fn query_param(raw_query: Option<&str>, key: &str) -> Option<String> {
-    let query = raw_query?;
-    for pair in query.split('&') {
-        let mut parts = pair.splitn(2, '=');
-        let k = parts.next().unwrap_or_default();
-        if k != key {
-            continue;
-        }
-        let value = parts.next().unwrap_or_default();
-        return decode_query_value(value).or_else(|| Some(value.to_string()));
-    }
-    None
-}
-
-fn video_id_from_embed_referer(request_headers: &HeaderMap) -> Option<String> {
-    let referer = request_headers.get(header::REFERER)?.to_str().ok()?;
-    let re = EMBED_REFERER_VIDEO_ID_RE.get_or_init(|| {
-        Regex::new(r"/api/youtube/embed/([A-Za-z0-9_-]{11})(?:[/?#]|$)")
-            .expect("valid embed referer video_id regex")
-    });
-    let captures = re.captures(referer)?;
-    let video_id = captures.get(1)?.as_str();
-    if is_valid_video_id(video_id) {
-        Some(video_id.to_string())
-    } else {
-        None
-    }
-}
-
-fn itag_quality_label(itag: &str) -> &'static str {
-    match itag {
-        "137" => "1080p (avc1)",
-        "248" => "1080p (vp9)",
-        "399" => "1080p (av1)",
-        "299" => "1080p60 (avc1)",
-        "136" => "720p (avc1)",
-        "247" => "720p (vp9)",
-        "398" => "720p (av1)",
-        "22" => "720p (avc1)",
-        "298" => "720p60",
-        "135" => "480p (avc1)",
-        "244" => "480p (vp9)",
-        "397" => "480p (av1)",
-        "18" => "360p (avc1)",
-        "134" => "360p (avc1)",
-        "243" => "360p (vp9)",
-        "396" => "360p (av1)",
-        "133" => "240p (avc1)",
-        "242" => "240p (vp9)",
-        "395" => "240p (av1)",
-        "160" => "144p (avc1)",
-        "278" => "144p (vp9)",
-        "394" => "144p (av1)",
-        _ => "unknown",
-    }
-}
-
-fn observe_quality_from_request(
-    state: &YoutubeState,
-    raw_query: Option<&str>,
-    request_headers: &HeaderMap,
-) {
-    let Some(itag) = query_param(raw_query, "itag") else {
-        return;
-    };
-    if !is_video_itag(&itag) {
-        return;
-    }
-    let video_id = if let Some(id) = video_id_from_embed_referer(request_headers) {
-        id
-    } else if let Some(raw_video_id) = query_param(raw_query, "id") {
-        let base = raw_video_id.split('.').next().unwrap_or(&raw_video_id);
-        if is_valid_video_id(base) {
-            base.to_string()
-        } else {
-            return;
-        }
-    } else {
-        return;
-    };
-
-    let mut observations = match state.quality_observations.lock() {
-        Ok(o) => o,
-        Err(_) => return,
-    };
-
-    let now = now_unix_secs();
-    observations.retain(|_, v| now.saturating_sub(v.last_updated_unix_secs) <= 3600);
-
-    let quality_label = itag_quality_label(&itag).to_string();
-    let obs = observations
-        .entry(video_id.clone())
-        .or_insert_with(|| QualityObservation {
-            current_itag: itag.clone(),
-            current_quality_label: quality_label.clone(),
-            counts: HashMap::new(),
-            last_updated_unix_secs: now,
-        });
-
-    obs.current_itag = itag.clone();
-    obs.current_quality_label = quality_label;
-    obs.last_updated_unix_secs = now;
-    *obs.counts.entry(itag).or_insert(0) += 1;
-
-    let snapshot = build_quality_observed_response(obs);
-    drop(observations);
-
-    let sender = get_or_create_quality_sender(state, &video_id);
-    let _ = sender.send(snapshot);
 }
 
 fn inject_quality_indicator_script(html: &str, video_id: &str) -> String {
@@ -1560,21 +1285,6 @@ mod tests {
         assert_eq!(
             rewrite_html_urls(input, "https://inv.wandabanet.de"),
             expected
-        );
-    }
-
-    #[test]
-    fn video_id_from_embed_referer_extracts_video_id() {
-        let mut headers = HeaderMap::new();
-        headers.insert(
-            header::REFERER,
-            HeaderValue::from_static(
-                "http://localhost:8080/api/youtube/embed/vYy4em2fQ8Q?autoplay=1",
-            ),
-        );
-        assert_eq!(
-            video_id_from_embed_referer(&headers),
-            Some("vYy4em2fQ8Q".to_string())
         );
     }
